README
Barangay Clearance City Wide Information System.
Barangay Clearance or also known as Certificate of Residency is a government-issued document in the Philippines. Usually, this certificate is used for applying business permit. This includes but not limited to a requirement for Local Employment, Government Agency Transactions, Travel Abroad, Police Clearance, NBI Clearance, Loan Application, Scholarship Application and several others. People can get a certificate on their barangay and also in the City Hall or Municipal Hall. Barangay Clearance is a document certifying that the applicant is of good moral character and law-abiding and good resident of a given town or barangay. It also proves that the applicant has no bad records or immoral background.